% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_MCMC_compute_psm.R
\name{compute_psm}
\alias{compute_psm}
\title{Compute and Plot the Posterior Similarity Matrix}
\usage{
compute_psm(object, group_specific = FALSE, plot = TRUE, ncores = 0)
}
\arguments{
\item{object}{An object of class \code{SANmcmc}.}

\item{group_specific}{Logical (default \code{FALSE}). If \code{group_specific = FALSE}, the function considers the overall PSM.
If \code{group_specific = TRUE}, the function considers the group-specific PSMs.}

\item{plot}{Logical (default \code{TRUE}). Whether to plot the PSM.}

\item{ncores}{A parameter to pass to the \code{salso::salso()} function (only for \code{SANmcmc} objects). The number of CPU cores to use for parallel computing; a value of zero indicates the use of all cores on the system.}
}
\value{
The function \code{compute_psm} returns and plots the posterior similarity matrix. If \code{group_specific = FALSE},
the output is a matrix of dimension N x N. If \code{group_specific = TRUE}, the output is a list on length G, where each
entry contains a matrix of dimension Nj x Nj.
}
\description{
The function computes and plots the posterior similarity matrix (PSM), either for the whole dataset, or separately for each group.
The function takes as input an object from \code{fit_CAM}, \code{fit_fiSAN}, or \code{fit_fSAN},
used with the \code{est_method = "MCMC"} argument.
}
\examples{
# Generate example data
set.seed(123)
y <- c(rnorm(100),rnorm(100,5))
g <- rep(1:2,rep(100,2))
plot(y,col=g)
# Fitting fiSAN via MCMC
est <- fit_fiSAN(y, g, est_method = "MCMC")
est
# Estimate PSM
psm_overall <- compute_psm(est)
# Estimate group-specific PSMs
psm_gs <- compute_psm(est, group_specific =  TRUE)

}
