% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_VI_estimate_G.R
\name{estimate_G}
\alias{estimate_G}
\alias{plot.SANvi_G}
\alias{print.SANvi_G}
\title{Estimate the Posterior Atoms and Weights of the Discrete Mixing Distributions}
\usage{
estimate_G(object, ...)

\method{plot}{SANvi_G}(x, DC_num = NULL, lim = 2, ...)

\method{print}{SANvi_G}(x, thr = 0.01, ...)
}
\arguments{
\item{object}{an object of class \code{SANvi}.}

\item{...}{ignored.}

\item{x}{an object of class \code{summary_SANvi}.}

\item{DC_num}{an integer or a vector of integers indicating which distributional clusters to plot.}

\item{lim}{optional value for \code{plot} method to adjust the limits of the x-axis. The atoms are plotted on a range
given by \code{min(posterior means)-2, max(posterior means)+2}. The default is set to 2.}

\item{thr}{argument for the \code{print()} method. It should be a small positive number,
representing a threshold. If the posterior weight of a shared atom is below the threshold, the
atom is not reported.}
}
\value{
an object of class \code{summary_SANvi}, which is a matrix comprising posterior means,
variances, and a column for each estimated DC containing the posterior weights.
}
\description{
This function estimates the posterior atoms and weights characterizing the discrete
mixing distributions using the variational estimates obtained from one of the models implemented in SANvi.
}
\examples{
# Generate example data
set.seed(1232)
y <- c(rnorm(100),rnorm(100,5))
g <- rep(1:2,rep(100,2))

# Fitting fiSAN via variational inference
est <- fit_fiSAN(y,g)

# Estimate posterior atoms and weights
estimate_G(est)
}
