% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_getfunctions.R
\name{get_model}
\alias{get_model}
\alias{get_model.SANvi}
\alias{get_model.SANmcmc}
\alias{get_time}
\alias{get_time.SANvi}
\alias{get_time.SANmcmc}
\alias{get_params}
\alias{get_params.SANvi}
\alias{get_params.SANmcmc}
\alias{get_sim}
\alias{get_sim.SANvi}
\alias{get_sim.SANmcmc}
\title{Accessor functions for SAN model output}
\usage{
get_model(object)

\method{get_model}{SANvi}(object)

\method{get_model}{SANmcmc}(object)

get_time(object)

\method{get_time}{SANvi}(object)

\method{get_time}{SANmcmc}(object)

get_params(object)

\method{get_params}{SANvi}(object)

\method{get_params}{SANmcmc}(object)

get_sim(object)

\method{get_sim}{SANvi}(object)

\method{get_sim}{SANmcmc}(object)
}
\arguments{
\item{object}{An object of class \code{SANmcmc} or \code{SANvi}.}
}
\value{
A component from the model object (see details).
}
\description{
\code{get_model}, \code{get_time}, \code{get_params}, and \code{get_sim} are accessor functions that extract key components from objects of class \code{SANmcmc} or \code{SANvi},
depending on whether the model was estimated via MCMC or variational inference. In particular:

\itemize{
  \item \code{get_model}: returns a character string indicating the model type;
  \item \code{get_time}: returns the runtime of the estimation procedure;
  \item \code{get_params}: returns the list of data and prior hyperparameters;
  \item \code{get_sim}: returns the fitted quantities (e.g., MCMC or VI estimates).
}
}
\examples{
set.seed(123)
y <- c(rnorm(40, 0, 0.3), rnorm(20, 5, 0.3))
g <- c(rep(1:6, each = 10))
out <- fit_fSAN(y = y, group = g, est_method = "MCMC",
                mcmc_param = list(nrep = 500, burn = 200))
get_model(out)
get_time(out)
hp <- get_params(out)
sims <- get_sim(out)

}
