% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_SANBA_plotting.R
\name{plot.SANvi}
\alias{plot.SANvi}
\title{Plot the variational inference output}
\usage{
\method{plot}{SANvi}(x, ...)
}
\arguments{
\item{x}{object of class \code{SANvi}.}

\item{...}{additional graphical parameters to be passed.}
}
\value{
The function plots the path followed by the ELBO and its subsequent differences.
}
\description{
Plot method for objects of class \code{SANvi}.
The function displays two graphs, showing the behavior of the ELBO values as
a function of the iterations and its subsequent increments, which should always be positive (log-scale).
}
