% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SANBA_printing.R
\name{print.SANmcmc}
\alias{print.SANmcmc}
\title{Print the MCMC Output}
\usage{
\method{print}{SANmcmc}(x, ...)
}
\arguments{
\item{x}{Object of class \code{SANmcmc}.}

\item{...}{Ignored.}
}
\value{
The function prints a summary of the fitted model.
}
\description{
Print method for objects of class \code{SANmcmc}.
}
