% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SANBA_printing.R
\name{print.SANvi}
\alias{print.SANvi}
\title{Print the variational inference output}
\usage{
\method{print}{SANvi}(x, ...)
}
\arguments{
\item{x}{object of class \code{SANvi}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function prints a summary of the fitted model.
}
\description{
Print method for objects of class \code{SANvi}.
}
