% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SANba.R
\docType{package}
\name{SANba-package}
\alias{SANba}
\alias{SANba-package}
\title{SANba: A Meta-Package for Fitting Common and Shared Atoms Nested Models with MCMC and Variational Inference}
\description{
An R meta-package calling R packages for estimating Bayesian nested mixture models via MCMC and Variational Bayes methods. Specifically, the package implements the common atoms model (Denti et al., 2023), its finite version (D'Angelo et al., 2023), and a hybrid finite-infinite model. All models use Gaussian mixtures with a normal-inverse-gamma prior distribution on the parameters. Additional functions are provided to help analyzing the results of the fitting procedure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fradenti/SANba}
  \item Report bugs at \url{https://github.com/fradenti/SANba/issues}
}

}
\author{
\strong{Maintainer}: Francesco Denti \email{francescodenti.personal@gmail.com} (\href{https://orcid.org/0000-0003-2978-4702}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Laura D'Angelo \email{laura.dangelo@live.com} (\href{https://orcid.org/0000-0001-5034-7414}{ORCID}) [copyright holder]
}

}
\keyword{internal}
