% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_SANBA_summary_estimate_partitions.R
\name{summary}
\alias{summary}
\alias{summary.SANvi}
\alias{summary.SANmcmc}
\alias{print.summary_mcmc}
\alias{print.summary_vi}
\alias{plot.summary_mcmc}
\alias{plot.summary_vi}
\title{Summarize the results estimating observational and distributional clusters}
\usage{
\method{summary}{SANvi}(object, ordered = TRUE, ...)

\method{summary}{SANmcmc}(object, ordered = TRUE, add_burnin = 0, ncores = 0, ...)

\method{print}{summary_mcmc}(x, ...)

\method{print}{summary_vi}(x, ...)

\method{plot}{summary_mcmc}(
  x,
  DC_num = NULL,
  type = c("ecdf", "boxplot", "scatter"),
  palette_brewed = FALSE,
  ...
)

\method{plot}{summary_vi}(
  x,
  DC_num = NULL,
  type = c("ecdf", "boxplot", "scatter"),
  palette_brewed = FALSE,
  ...
)
}
\arguments{
\item{object}{object of class \code{SANmcmc} (the result of a call to \code{\link{fit_fiSAN}},
\code{\link{fit_fSAN}}, or \code{\link{fit_CAM}} with \code{method = "MCMC"}) or \code{SANvi}
(the result of a call to \code{\link{fit_fiSAN}},\code{\link{fit_fSAN}}, or \code{\link{fit_CAM}} with \code{method = "VI"}).}

\item{ordered}{logical, if \code{TRUE} (default), the function sorts the distributional cluster labels reflecting the
increasing values of medians of the data assigned to each DC.}

\item{...}{ignored.}

\item{add_burnin}{integer (default = 0). Optional number of observations to additionally discard.}

\item{ncores}{for MCMC objects, a parameter of the function \code{salso::salso()}. The number of CPU cores to use, i.e., the number of simultaneous runs at any given time.
A value of zero indicates the use of all cores on the system.}

\item{x}{an object of class \code{variational_estimates}, which can be obtained from the
function \code{\link{summary}}.}

\item{DC_num}{an integer or a vector of integers indicating which distributional clusters to plot.}

\item{type}{what type of plot should be drawn (only for the left-side plot). Possible types are "boxplot", "ecdf", and "scatter".}

\item{palette_brewed}{(logical) the color palette to be used. Default is \code{R} base colors (\code{palette_brewed = FALSE}).}
}
\value{
a list of class \code{summary_vi} or \code{summary_mcmc} containing
\itemize{
  \item \code{obs_level}: a data frame containing the data values, their group indexes, and the observational and distributional clustering assignments for each observation.
  \item \code{dis_level}: a vector with the distributional clustering assignment for each unit.
}
}
\description{
Given the output of a \code{sanba} function, estimate the observational and distributional partitions using \code{\link[salso:salso]{salso::salso()}} for MCMC, and maximum a posteriori for VI.
}
\examples{
set.seed(123)
## ....

}
\seealso{
\code{\link[salso:salso]{salso::salso()}}, \code{\link{print.SANmcmc}}, \code{\link{plot.SANmcmc}}
}
