% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SANBA_summary.R
\name{summary.SANmcmc}
\alias{summary.SANmcmc}
\title{Summarize the MCMC Output}
\usage{
\method{summary}{SANmcmc}(object, ...)
}
\arguments{
\item{object}{of class \code{SANmcmc}.}

\item{...}{Ignored.}
}
\value{
The function prints a summary of the fitted model.
}
\description{
Summary method for objects of class \code{SANmcmc}.
}
