% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SANBA_summary.R
\name{summary.SANvi}
\alias{summary.SANvi}
\title{Summarize the Variational Inference Output}
\usage{
\method{summary}{SANvi}(object, ...)
}
\arguments{
\item{object}{Object of class \code{SANvi}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function prints a summary of the fitted model.
}
\description{
Summary method for objects of class \code{SANvi}.
}
