% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanba-package.R
\docType{package}
\name{sanba-package}
\alias{sanba}
\alias{sanba-package}
\title{sanba: Fitting Shared Atoms Nested Models via MCMC or Variational Bayes}
\description{
An efficient tool for fitting nested mixture models based on a shared set of atoms via Markov Chain Monte Carlo and variational inference algorithms. Specifically, the package implements the common atoms model (Denti et al., 2023), its finite version (similar to D'Angelo et al., 2023), and a hybrid finite-infinite model (D'Angelo and Denti, 2024). All models implement univariate nested mixtures with Gaussian kernels equipped with a normal-inverse gamma prior distribution on the parameters. Additional functions are provided to help analyze the results of the fitting procedure. References: Denti, Camerlenghi, Guindani, Mira (2023) \doi{10.1080/01621459.2021.1933499}, D’Angelo, Canale, Yu, Guindani (2023) \doi{10.1111/biom.13626}, D’Angelo, Denti (2024) \doi{10.1214/24-BA1458}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fradenti/sanba}
  \item Report bugs at \url{https://github.com/fradenti/sanba/issues}
}

}
\author{
\strong{Maintainer}: Francesco Denti \email{francescodenti.personal@gmail.com} (\href{https://orcid.org/0000-0001-5034-7414}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Laura D'Angelo \email{laura.dangelo@live.com} (\href{https://orcid.org/0000-0003-2978-4702}{ORCID})
}

}
\keyword{internal}
